requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib